requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.7.2 ; python_version >= '3.7'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
bs4==0.0.1
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
contourpy==1.1.0 ; python_version >= '3.8'
crispy-bootstrap5==0.7
cycler==0.11.0 ; python_version >= '3.6'
django==4.1
django-cleanup==8.0.0
django-crispy-forms==2.0
django-htmx==1.16.0
exceptiongroup==1.1.3 ; python_version < '3.11'
fonttools==4.42.1 ; python_version >= '3.8'
frozenlist==1.4.0 ; python_version >= '3.8'
idna==3.4 ; python_version >= '3.5'
importlib-resources==6.0.1 ; python_version < '3.10'
iniconfig==2.0.0 ; python_version >= '3.7'
joblib==1.3.2 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
matplotlib==3.7.2
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.25.2 ; python_version >= '3.9'
openai==0.27.8
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.3
pillow==10.0.0 ; python_version >= '3.8'
plotly==5.16.1
pluggy==1.2.0 ; python_version >= '3.7'
pyarrow==12.0.1
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pypdf2==3.0.1
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-decouple==3.8
pytz==2023.3
regex==2023.8.8 ; python_version >= '3.6'
reportlab==4.0.4
requests==2.31.0 ; python_version >= '3.7'
scikit-learn==1.3.0
scipy==1.11.2 ; python_version < '3.13' and python_version >= '3.9'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.4.1 ; python_version >= '3.7'
sqlparse==0.4.4 ; python_version >= '3.5'
tenacity==8.2.3 ; python_version >= '3.7'
threadpoolctl==3.2.0 ; python_version >= '3.8'
tiktoken==0.4.0
tomli==2.0.1 ; python_version < '3.11'
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.7.1 ; python_version < '3.10'
tzdata==2023.3 ; python_version >= '2'
urllib3==2.0.4 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
youtube-transcript-api==0.6.1
zipp==3.16.2 ; python_version < '3.10'
